Moved LicenseGroup bindings into new LicenseGroupService.
